% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{assignDfToDict}
\alias{assignDfToDict}
\title{assignDfToDict}
\usage{
assignDfToDict(df)
}
\arguments{
\item{df}{the dataframe in question}
}
\value{
Void
}
\description{
assigns an empty metadata list to the data dictionary for the dataframe in question
}
