% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_extract.R
\name{constCharLength}
\alias{constCharLength}
\title{Does the variable have a constant length}
\usage{
constCharLength(vect)
}
\arguments{
\item{vect}{the vector in question}
}
\description{
Does the variable have a constant length
}
