% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_extract.R
\name{dateForm}
\alias{dateForm}
\title{Provisional date format checking function}
\usage{
dateForm(vect)
}
\arguments{
\item{vect}{the vector in question}
}
\value{

}
\description{
Provisional date format checking function
}
