% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{getConnectedMeta}
\alias{getConnectedMeta}
\title{getConnectedMeta}
\usage{
getConnectedMeta(df, colname)
}
\arguments{
\item{df}{The Dataframe in question}

\item{colname}{The column name in question}
}
\value{
the associated metadata
}
\description{
Finds the metadata associated with the specified dataframe and column
}
