% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{getDataDict}
\alias{getDataDict}
\title{getDataDict}
\usage{
getDataDict()
}
\value{
data_dict - Either the existing data dictionary or a new one if it is missing
}
\description{
Fetches the data_dict list, or creates it if it is missing
}
