% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_extract.R
\name{guessDataType}
\alias{guessDataType}
\title{Guess the data type based on the attributes}
\usage{
guessDataType(vect)
}
\arguments{
\item{vect}{the vector in question}
}
\description{
Guess the data type based on the attributes
}
