% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{makeEmptyMeta}
\alias{makeEmptyMeta}
\title{makeEmptyMeta}
\usage{
makeEmptyMeta(df)
}
\arguments{
\item{df}{the dataframe in question}
}
\value{
the empty named list
}
\description{
Makes an empty metadata object
}
