% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{package}
\name{megametadata}
\alias{megametadata}
\alias{meta-monster}
\alias{megametadata-package}
\title{Data Dictionary generation and interaction}
\description{
This package contains functions and examples for using and expanding data
dictionaries
}
