% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yaml_integrations.R
\name{meta_AutoClassifier}
\alias{meta_AutoClassifier}
\title{process to append the typing of a column at read time to the meta_dict}
\usage{
meta_AutoClassifier(column, meta_list, dataset)
}
\arguments{
\item{column}{The column to type}

\item{meta_list}{A dictionary for the dataset}

\item{dataset}{The dataset to be documenting}
}
\description{
process to append the typing of a column at read time to the meta_dict
}
