% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yaml_integrations.R
\name{meta_AutoTyper}
\alias{meta_AutoTyper}
\title{Process to guess the data category of the columns for the meta_dict}
\usage{
meta_AutoTyper(column, meta_list, dataset)
}
\arguments{
\item{column}{The column to type}

\item{meta_list}{A dictionary for the dataset}

\item{dataset}{The dataset to be documenting}
}
\description{
Process to guess the data category of the columns for the meta_dict
}
