% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yaml_integrations.R
\name{meta_Colnamer}
\alias{meta_Colnamer}
\title{Process to name the columns in the meta_dict}
\usage{
meta_Colnamer(column, meta_list)
}
\arguments{
\item{column}{The column to type}

\item{meta_list}{A dictionary for the dataset}
}
\description{
Process to name the columns in the meta_dict
}
