% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_extract.R
\name{meta_GuessDataType}
\alias{meta_GuessDataType}
\title{Guess the data type based on the attributes}
\usage{
meta_GuessDataType(vect)
}
\arguments{
\item{vect}{the vector in question}
}
\description{
Guess the data type based on the attributes
}
