% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yaml_integrations.R
\name{meta_UpdateDictWithSpec}
\alias{meta_UpdateDictWithSpec}
\title{Update dictionary to fit specification}
\usage{
meta_UpdateDictWithSpec(a_dict, a_spec, a_dataset, a_level)
}
\arguments{
\item{a_dict}{A dictionary for the dataset}

\item{a_spec}{The specification for the dictionary}

\item{a_dataset}{The dataset to be documenting}

\item{a_level}{The level to perform the documentation}
}
\description{
Function which takes an existing dictionary template and returns an updated
   version which fits the specification.
}
