% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yaml_integrations.R
\name{metaAutoTyper}
\alias{metaAutoTyper}
\title{Process to guess the data category of the columns for the meta_dict}
\usage{
metaAutoTyper(column, meta_list, dataset)
}
\arguments{
\item{dataset}{}
}
\description{
Process to guess the data category of the columns for the meta_dict
}
