% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_functions.R
\name{metaClassExtract}
\alias{metaClassExtract}
\title{Class Extractor}
\usage{
metaClassExtract(slice, metadata, extract_level)
}
\arguments{
\item{slice}{the index of the data field}

\item{metadata}{the metadata dictionary}

\item{extract_level}{the level to perform the extract on}
}
\description{
Class Extractor
}
