% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yaml_integrations.R
\name{metaColnamer}
\alias{metaColnamer}
\title{Process to name the columns in the meta_dict}
\usage{
metaColnamer(column, meta_list)
}
\arguments{
\item{meta_list}{}
}
\description{
Process to name the columns in the meta_dict
}
