% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_setup.R
\name{metaDictionary.default}
\alias{metaDictionary.default}
\title{The default meta_dictionary generator}
\usage{
\method{metaDictionary}{default}(the_data, dict = list())
}
\arguments{
\item{the_data}{the data to be documented}

\item{dict}{the dictionary to append it to}
}
\description{
The default meta_dictionary generator
}
