% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_setup.R
\name{metaDictionary.list}
\alias{metaDictionary.list}
\title{Method for lists in the generic}
\usage{
\method{metaDictionary}{list}(the_data, dict = list, levelgen = "DefaultLevel")
}
\arguments{
\item{the_data}{the data to be documented}

\item{dict}{the dictionary to append it to}

\item{levelgen}{the level to generate the metadata at}
}
\value{

}
\description{
Method for lists in the generic
}
