% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_functions.R
\name{metaSplitNominal}
\alias{metaSplitNominal}
\title{Variable Type Splitter}
\usage{
metaSplitNominal(metadata, extract_level = NULL)
}
\arguments{
\item{metadata}{the data fictionary}

\item{extract_level}{the level to perform the extract on}
}
\value{

}
\description{
Variable Type Splitter
}
