% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yaml_integrations.R
\name{metaUpdateDictWithSpec}
\alias{metaUpdateDictWithSpec}
\title{Update dictionary to fit specification}
\usage{
metaUpdateDictWithSpec(a_dict, a_spec, a_dataset, a_level)
}
\arguments{
\item{a_level}{}
}
\description{
Function which takes an existing dictionary template and returns an updated
   version which fits the specification.
}
