% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{spec}
\alias{spec}
\title{Example Specification}
\format{A list of lists with each Level of the data and each class of column}
\usage{
spec
}
\description{
An example specification which can be used to expand or update the initial
phase of the data dictionary
}
\keyword{datasets}
