% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{tester}
\alias{tester}
\title{Example Dict config}
\format{
A list of lists with each Level of the data and each class of column
}
\usage{
tester
}
\description{
An example Dictionary which can be used as an example to expand from or as a
testing object
}
\keyword{datasets}
