% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_extract.R
\name{util_ConstCharLength}
\alias{util_ConstCharLength}
\title{Does the variable have a constant length}
\usage{
util_ConstCharLength(vect)
}
\arguments{
\item{vect}{the vector in question}
}
\description{
Does the variable have a constant length
}
