% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_extract.R
\name{util_DateForm}
\alias{util_DateForm}
\title{Provisional date format checking function}
\usage{
util_DateForm(vect)
}
\arguments{
\item{vect}{the vector in question}
}
\description{
Provisional date format checking function
}
