% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_extract.R
\name{util_Distinctiveness}
\alias{util_Distinctiveness}
\title{Function to compute the distinctiveness of a function, using a variant log-loss}
\usage{
util_Distinctiveness(vect)
}
\arguments{
\item{vect}{the vector in question}
}
\description{
Function to compute the distinctiveness of a function, using a variant log-loss
}
