% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yaml_integrations.R
\name{util_EvalOnData}
\alias{util_EvalOnData}
\title{Function to evaluate a string name of function on some specified data}
\usage{
util_EvalOnData(x, data)
}
\arguments{
\item{x}{the name of the function to evaluate on the data}

\item{data}{the data to evaluate the function on}
}
\description{
Function to evaluate a string name of function on some specified data
}
