% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_extract.R
\name{util_PostForm}
\alias{util_PostForm}
\title{Provisional postcode format checker}
\usage{
util_PostForm(vect)
}
\arguments{
\item{vect}{the vector in question}
}
\description{
Provisional postcode format checker
}
