% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yaml_integrations.R
\name{util_Range}
\alias{util_Range}
\title{Function to determine the range of an integer list}
\usage{
util_Range(x)
}
\arguments{
\item{x}{the vector to determine the range of}
}
\description{
Function to determine the range of an integer list
}
