% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_extract.R
\name{util_Uniqueness}
\alias{util_Uniqueness}
\title{Get the uniqueness of a vector}
\usage{
util_Uniqueness(vect)
}
\arguments{
\item{vect}{the vector in question}
}
\description{
Get the uniqueness of a vector
}
