% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yaml_integrations.R
\name{utilEvalOnData}
\alias{utilEvalOnData}
\title{Function to evaluate a string name of function on some specified data}
\usage{
utilEvalOnData(x, data)
}
\arguments{
\item{data}{}
}
\description{
Function to evaluate a string name of function on some specified data
}
