% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yaml_integrations.R
\name{utilRange}
\alias{utilRange}
\title{Function to determine the range of an integer list}
\usage{
utilRange(x)
}
\arguments{
\item{x}{the vector to determine the range of}
}
\value{

}
\description{
Function to determine the range of an integer list
}
