% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yaml_integrations.R
\name{metaAutoClassifier}
\alias{metaAutoClassifier}
\title{process to append the typing of a column at read time to the meta_dict}
\usage{
metaAutoClassifier(column, meta_list, dataset)
}
\arguments{
\item{dataset}{}
}
\description{
process to append the typing of a column at read time to the meta_dict
}
