% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_setup.R
\name{metaDictionary}
\alias{metaDictionary}
\title{Method dispatch for metadictionary}
\usage{
metaDictionary(the_data, dict)
}
\arguments{
\item{the_data}{the data to be documented}

\item{dict}{the dictionary to append it to}
}
\description{
Method dispatch for metadictionary
}
