% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yaml_integrations.R
\name{metaVariableAppend}
\alias{metaVariableAppend}
\title{Process to append the first pass variables to the meta_dict}
\usage{
metaVariableAppend(a_tibble, Leveled_meta = list())
}
\arguments{
\item{a_tibble}{The dataset to be documenting}

\item{Leveled_meta}{the level of the dataset dictionary to be documented}
}
\description{
Process to append the first pass variables to the meta_dict
}
